# Rewrite Buffer.BlockCopy in C#