--- a/src/System.Private.CoreLib/src/System/Runtime/CompilerServices/RuntimeHelpers.cs
+++ b/src/System.Private.CoreLib/src/System/Runtime/CompilerServices/RuntimeHelpers.cs
@@ -7,6 +7,13 @@ using System.Diagnostics.CodeAnalysis;
 using System.Runtime.InteropServices;
 using Internal.Runtime.CompilerServices;
 
+#pragma warning disable SA1121 // explicitly using type aliases instead of built-in types
+#if BIT64
+using nuint = System.UInt64;
+#else
+using nuint = System.UInt32;
+#endif
+
 namespace System.Runtime.CompilerServices
 {
     public static partial class RuntimeHelpers
@@ -164,12 +171,14 @@ namespace System.Runtime.CompilerServices
         internal static ref byte GetRawSzArrayData(this Array array) =>
             ref Unsafe.As<RawArrayData>(array).Data;
 
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        internal static unsafe ref byte GetRawArrayData(this Array array) =>
+            ref Unsafe.AddByteOffset(ref Unsafe.As<RawData>(array).Data, (nuint)GetObjectMethodTablePointer(array)->BaseSize - (nuint)(2 * sizeof(IntPtr)));
+
         internal static unsafe ushort GetElementSize(this Array array)
         {
             Debug.Assert(ObjectHasComponentSize(array));
-
-            // The first UINT16 of the method table is component size.
-            return *(ushort*)GetObjectMethodTablePointer(array);
+            return GetObjectMethodTablePointer(array)->ComponentSize;
         }
 
         // Returns true iff the object has a component size;
@@ -181,20 +190,30 @@ namespace System.Runtime.CompilerServices
             // [ pMethodTable || .. object data .. ]
             //   ^-- the object reference points here
             //
-            // The first DWORD of the method table class will have its high bit set if the
+            // The m_dwFlags field of the method table class will have its high bit set if the
             // method table has component size info stored somewhere. See member
             // MethodTable:IsStringOrArray in src\vm\methodtable.h for full details.
             //
             // So in effect this method is the equivalent of
             // return ((MethodTable*)(*obj))->IsStringOrArray();
+            return (int)GetObjectMethodTablePointer(obj)->Flags < 0;
+        }
 
-            Debug.Assert(obj != null);
-            return *(int*)GetObjectMethodTablePointer(obj) < 0;
+        // Subset of src\vm\methodtable.h
+        [StructLayout(LayoutKind.Explicit)]
+        private struct MethodTable
+        {
+            [FieldOffset(0)]
+            public ushort ComponentSize;
+            [FieldOffset(0)]
+            public uint Flags;
+            [FieldOffset(4)]
+            public uint BaseSize;
         }
 
         // Given an object reference, returns its MethodTable* as an IntPtr.
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static IntPtr GetObjectMethodTablePointer(object obj)
+        private static unsafe MethodTable* GetObjectMethodTablePointer(object obj)
         {
             Debug.Assert(obj != null);
 
@@ -202,7 +221,7 @@ namespace System.Runtime.CompilerServices
             // method table pointer, so just back up one pointer and immediately deref.
             // This is not ideal in terms of minimizing instruction count but is the best we can do at the moment.
 
-            return Unsafe.Add(ref Unsafe.As<byte, IntPtr>(ref obj.GetRawData()), -1);
+            return (MethodTable *)Unsafe.Add(ref Unsafe.As<byte, IntPtr>(ref obj.GetRawData()), -1);
 
             // The JIT currently implements this as:
             // lea tmp, [rax + 8h] ; assume rax contains the object reference, tmp is type IntPtr&