… src/inc/jithelpers.h
… src/vm/comutilnative.cpp