Fixes #27106