@jkotas 2019-10-16
Performance results:

Before:

|        Method | numElements |       Mean |
|-------------- |------------ |-----------:|
| CallBlockCopy |          10 |   8.049 ns |
| CallBlockCopy |         100 |  10.297 ns |
| CallBlockCopy |        1000 |  30.870 ns |
| CallBlockCopy |       10000 | 220.712 ns |

After:

|        Method | numElements |       Mean |
|-------------- |------------ |-----------:|
| CallBlockCopy |          10 |   8.862 ns |
| CallBlockCopy |         100 |  11.137 ns |
| CallBlockCopy |        1000 |  41.078 ns |
| CallBlockCopy |       10000 | 227.914 ns |

The managed implementation is slightly slower on average (<5%), but it is GC pause friendly. Achieving the GC pause friendliness in the unmanaged implementation would add more significant overhead.

The larger delta for 1000 elements has to do with tuning of the managed Memcpy implementation that apparently performs worse for this specific block size on my machine.