@benaadams 2019-10-16
> The managed implementation is slightly slower on average (<5%)

Memmove has scope to be improved; its setup for ymm sizes however it will only ever compile to xmm as it was implemented pre-intrinsics. 

May be worth switching `Block16` for `Vector128` and `Block64` for 2x `Vector256`? However, beyond that; probably not worth chasing down the rabbit hole for 5% as we've been here before several times and its very time consuming.