@benaadams 2019-10-16
> Maybe seeing if switching Block16 for Vector128 and Block64 for 2x Vector256 gives an improvement?

Even have a commit for it 😄  https://github.com/benaadams/coreclr/commit/5b90be64652a803f060ed49006cbbe44b2703adf