@jkotas 2019-10-16
> benaadams@5b90be6

That was https://github.com/dotnet/coreclr/pull/25172 . Trying to hand-tune memmove in software is a losing battle. I hope that hardware will add memmove instruction one day that will be superior to hand tuned implementations in every dimension.