@benaadams 2019-10-16
> > benaadams/coreclr@5b90be6

> That was #25172 .

😄 so it was; thought it strange I'd not done anything with it!

>  I hope that hardware will add memmove instruction one day that will be superior to hand tuned implementations in every dimension.

Agreed. its such a common operation; I still don't understand why isn't a simple instruction that then can dispatch to the memory or DMA controllers to do the work if the sizes makes sense, rather than have to write a CPU loop in software; and the the cpu can get on other other work or sleep...