@benaadams 2019-10-16
Compared to AVX; ERMSB isn't great below 128bytes and its +2% win starts at >= 2048 bytes

![image](https://user-images.githubusercontent.com/1142958/66924183-1269f900-f022-11e9-96f8-69fb21f5c2a5.png)


Also its sounds more sensitive to misalignment than SIMD (+20%)

![image](https://user-images.githubusercontent.com/1142958/66924237-2b72aa00-f022-11e9-832f-9719e73b8046.png)