@benaadams 2019-10-16
Saying that... I would be interested in a `Cpuid.IsGenuineIntel` test as some of the Ryzen latencies aren't great (e.g. `_pdep_u32` and `_pext_u32`)