@jkotas 2019-10-17
> Would this be a 3.1 (LTS) candidate?

Unlikely