@AntonLapounov 2019-10-18
> slightly slower on average (<5%), but it is GC pause friendly

For 10K elements we fall back to the non-friendly PInvoke: https://github.com/jkotas/coreclr/blob/managed-blockcopy/src/System.Private.CoreLib/shared/System/Buffer.cs#L468-L471
so it seems the actual performance hit is greater than 5%.  Try disabling that fallback.