@AntonLapounov 2019-10-18
> The managed implementation is slightly slower on average (<5%), but it is GC pause friendly.

(Corrected the question.) For 10K block size we fall back to a PInvoke, and for 10 to 1K block sizes the performance hit of the managed implementation seems more than 5%.  Is the 1K size indeed somehow exceptional or is that actually evidence that we'd better lower `MemmoveNativeThreshold` to fall back to the PInvoke earlier (at least for your particular machine and the current managed implementation that does not take advantage of intrinsics)?