@jkotas 2019-10-18
> 10 to 1K block sizes the performance hit of the managed implementation seems more than 5%

Yeah, you are right. The difference can be up to 20% around the MemmoveNativeThreshold. This is memory copy implementation is used in many other places in managed code. If there is a good reason to look at tuning it again, I would rather do it in a separate change.

> we'd better lower MemmoveNativeThreshold to fall back to the PInvoke earlier 

The current MemmoveNativeThreshold is 2k. The PInvoke overhead is still significant for this buffer size. On my machine, there is actually a performance drop as we cross this threshold and fallback to PInvoke. This would call for making this threshold higher.