@Robird 2020-11-12
> > [benaadams/coreclr@5b90be6](https://github.com/benaadams/coreclr/commit/5b90be6)
> 
> That was #25172 . Trying to hand-tune memmove in software is a losing battle. I hope that hardware will add memmove instruction one day that will be superior to hand tuned implementations in every dimension.

I think, acutly there is hardware DMA on broad platforms, rather than CPU register and load/store instructions.